Fix config loarder after moving it to internal package
