+++ cmd/rest-api/application/application.go
@@ -2,17 +2,17 @@ package application
 
 import (
 	"context"
-	"fmt"
 	"net/http"
 
 	"github.com/labstack/echo"
+	log "github.com/sirupsen/logrus"
 	"go.mongodb.org/mongo-driver/mongo"
 
-	"github.com/vsmoraes/open-ledger/factory"
+	"github.com/vsmoraes/open-ledger/internal/factory"
+	"github.com/vsmoraes/open-ledger/internal/storage"
 	"github.com/vsmoraes/open-ledger/ledger"
 	"github.com/vsmoraes/open-ledger/ledger/account"
 	"github.com/vsmoraes/open-ledger/ledger/movement"
-	"github.com/vsmoraes/open-ledger/storage"
 )
 
 type Application struct {
@@ -57,10 +57,13 @@ func (app *Application) Start(port string) {
 	rg.POST("", createMovementController(app.l))
 	rg.GET("", findMomentsController(app.mf))
 
+	log.WithFields(log.Fields{
+		"port": port,
+	}).Info("Starting HTTP server")
 	app.e.Logger.Fatal(app.e.Start(port))
 }
 
 func (app *Application) stop() {
-	fmt.Println("disconnecting...")
+	log.Info("Disconnecting...")
 	app.mc.Disconnect(context.Background())
 }
